HappyFox/requirements.txt
google-api-python-client==2.127.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
psycopg2-binary==2.9.9